deleted class QuizService and add its methods to QuizGenerator
