--- src/test/java/ru/otus/QuizServiceTest.java
@@ -1,12 +1,0 @@
-package ru.otus;
-
-import junit.framework.TestCase;
-
-public class QuizServiceTest extends TestCase {
-
-    public void testResultOfGame() {
-    }
-
-    public void testSetPlayerName() {
-    }
-}
